agent/requirements.txt
requests~=2.32.5
undetected-chromedriver
selenium
setuptools
pillow
fastapi
uvicorn[standard]
